Require axios and base-64 to run in browsers
Currently the node-gogs-client library only executes in node.js
interpreter since it need the native modules like http and buffer. I have
tried to install polyfills for the browser but not all needed libraries
are easy to install.

Replacing the http/https and buffer modules with axios and base-64 make
this library cross platform and simplify the lib/request.js code, with the
drawback of increasing the dependency tree (6 packages).